Extract BaseHIDHost for shared UHID functionality
Create base_host.py with common HID host functionality:
- Transport specification handling
- UHID device creation and cleanup
- Connection state management

UnifiedHIDHost now inherits from BaseHIDHost, removing
~60 lines of duplicated code.
